NGS2: pull up GemsGuideStarSearchOptions.CatalogChoice
